refactor: TableRepository 생성 및 로직 이동
