+++ src/main/java/kitchenpos/application/TableService.java
@@ -3,11 +3,10 @@ package kitchenpos.application;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.dao.OrderDao;
-import kitchenpos.dao.OrderTableDao;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.exception.NotFoundOrderTableException;
 import kitchenpos.exception.OrderTableConvertEmptyStatusException;
+import kitchenpos.repository.TableRepository;
 import kitchenpos.ui.dto.request.OrderTableChangeEmptyRequest;
 import kitchenpos.ui.dto.request.OrderTableChangeNumberOfGuestsRequest;
 import kitchenpos.ui.dto.request.OrderTableCreateRequest;
@@ -16,12 +15,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class TableService {
+    private final TableRepository tableRepository;
     private final OrderDao orderDao;
-    private final OrderTableDao orderTableDao;
 
-    public TableService(OrderDao orderDao, OrderTableDao orderTableDao) {
+    public TableService(OrderDao orderDao, TableRepository tableRepository) {
         this.orderDao = orderDao;
-        this.orderTableDao = orderTableDao;
+        this.tableRepository = tableRepository;
     }
 
     @Transactional
@@ -29,21 +28,18 @@ public class TableService {
         OrderTable orderTable =
                 new OrderTable(orderTableCreateRequest.getNumberOfGuests(), orderTableCreateRequest.getEmpty());
 
-        return orderTableDao.save(orderTable);
+        return tableRepository.save(orderTable);
     }
 
     public List<OrderTable> list() {
-        return orderTableDao.findAll();
+        return tableRepository.findAll();
     }
 
     @Transactional
     public OrderTable changeEmpty(Long orderTableId, OrderTableChangeEmptyRequest orderTableChangeEmptyRequest) {
         validateOrderTableStatus(orderTableId);
 
-        OrderTable savedOrderTable = findOrderTable(orderTableId);
-        savedOrderTable.changeEmpty(orderTableChangeEmptyRequest.getEmpty());
-
-        return orderTableDao.save(savedOrderTable);
+        return tableRepository.changeEmpty(orderTableId, orderTableChangeEmptyRequest.getEmpty());
     }
 
     private void validateOrderTableStatus(Long orderTableId) {
@@ -56,14 +52,7 @@ public class TableService {
     @Transactional
     public OrderTable changeNumberOfGuests(Long orderTableId,
                                            OrderTableChangeNumberOfGuestsRequest orderTableChangeNumberOfGuestsRequest) {
-        OrderTable savedOrderTable = findOrderTable(orderTableId);
-        savedOrderTable.changeNumberOfGuests(orderTableChangeNumberOfGuestsRequest.getNumberOfGuests());
-
-        return orderTableDao.save(savedOrderTable);
-    }
-
-    private OrderTable findOrderTable(Long orderTableId) {
-        return orderTableDao.findById(orderTableId)
-                .orElseThrow(NotFoundOrderTableException::new);
+        return tableRepository.changeNumberOfGuests(orderTableId,
+                orderTableChangeNumberOfGuestsRequest.getNumberOfGuests());
     }
 }
